+++ CHANGELOG.md
@@ -1,31 +1,45 @@
+* [20.02.2022]
+  * service worker caching optimzied
+  * new news toggle with iframes
+  * shorten changelog
+  
 * [18.02.2022]
   * store refactoring
   * product flags isNew, isHot, isNewSoon now in store
+  
 * [17.02.2022]
   * solved great history refactoring
   * remove deprecated dependencies
   * update the inst.json
   * fix dart-sass changes
   * no future arrow right vor today changes
+  
 * [16.02.2022]
   * new partNr for many products
+  
 * [12.02.2022]
   * prepare notifications
+  
 * [10.02.2022]
   * multiple heartlist
+  
 * [08.02.2022]
   * prettier implemantation
   * fix safari datepicker
+  
 * [07.02.2022]
   * add partNr of bricklink/lego
+  
 * [04.02.2022]
   * fix firefox zoom bug for specials burgblaustein
   * fix tooltip width on minimobile + other mobile fixxes
   * extract footer
   * fix some url bugs 
   * prepare notifications for heartlist changes
+  
 * [03.02.2022]
   * add cherrio to parse pages, speed up 14sek to 2sek
+  
 * [02.02.2022]
   * add new fetchnparse code
   * solve many history issues
@@ -33,37 +47,49 @@
   * fix some tag issues
   * add fetch interval for changes
   * new welcome page with support
+  
 * [27.01.2022]
   * add time ago into product history
   * active product in url
+  
 * [26.01.2022]
   * switch in brickbar german title to english
   * tags are better visile in darkmode now
+  
 * [24.01.2022]
   * extract history to single file
+  
 * [21.01.2022]
   * prepare partner
   * add newest tags for newest announced products
+  
 * [20.01.2022]
   * add affiliate notice
+  
 * [17.01.2022]
   * add two movie names
   * change `parts` to `brickbar`
   * add `american` products to `usa` tag
+  
 * [15.01.2022]
   * podcast interview
+  
 * [14.01.2022]
   * switch from `#` hash to `?` search url's
   * add to tooltip `1 Std. her` or `2 Tage her` - time ago
   * update inst.json
   * add BurgBlaustein easteregg
   * refactore eastereggs
+  
 * [13.01.2022]
   * add scrollIntoView to Tooltip
+  
 * [11.01.2022]
   * add Netherland easteregg
+  
 * [08.01.2022]
   * Manhattan zoom bug solved and add tooltip to map https://bbx.watch/#tags=manhattan
+  
 * [09.01.2022]
   * add Darkmode toggle with Animation
   * add https://bbx.watch/#tags=bundeswehr Tag and 3 new Xingbao BB Pro Products